requirements.txt
omegaconf==2.3.0
opencv-python==4.8.0.76       # import cv2
Pillow==9.5.0                 # import PIL / from PIL import Image
scikit-image==0.21.0          # from skimage import io
scipy==1.11.1                 # from scipy.spatial.transform import Rotation
cmake==3.27.7
face_recognition==1.3.0       # import face_recognition
facenet-pytorch==2.5.3        # from facenet_pytorch import MTCNN, InceptionResnetV1
ptgaze==0.2.3                 # from ptgaze.common.face_model import ...
PyYAML==6.0.1                 # import yaml
glob2==0.7                    # import glob2 as gb
timm==0.9.2
mediapipe==0.10.14
pandas==2.0.3
numpy==1.26.4